Revert "QuickAdminPanel file update"

+++ database/migrations/2020_10_13_000022_add_relationship_fields_to_imageslides_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToImageslidesTable extends Migration
     public function up()
     {
         Schema::table('imageslides', function (Blueprint $table) {
-            $table->unsignedBigInteger('page_id')->nullable();
+            $table->unsignedInteger('page_id')->nullable();
             $table->foreign('page_id', 'page_fk_2180059')->references('id')->on('pages');
         });
     }
