+++ database/migrations/2020_10_13_000024_add_relationship_fields_to_ministries_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToMinistriesTable extends Migration
     public function up()
     {
         Schema::table('ministries', function (Blueprint $table) {
-            $table->unsignedBigInteger('authority_id')->nullable();
+            $table->unsignedInteger('authority_id')->nullable();
             $table->foreign('authority_id', 'authority_fk_2164519')->references('id')->on('roles');
         });
     }
